[ci skip] Improved readability.

+++ src/query/execute_statement.h
@@ -190,8 +190,8 @@ extern int do_set_timezone (PARSER_CONTEXT * parser, PT_NODE * statement);
 extern int do_set_query_trace (PARSER_CONTEXT * parser, PT_NODE * statement);
 extern int do_kill (PARSER_CONTEXT * parser, PT_NODE * statement);
 
-extern int do_find_class_by_query (const char *name, char *buf, size_t buf_size);
-extern int do_find_serial_by_query (const char *name, char *buf, size_t buf_size);
-extern int do_find_trigger_by_query (const char *name, char *buf, size_t buf_size);
+extern int do_find_class_by_query (const char *name, char *buf, int buf_size);
+extern int do_find_serial_by_query (const char *name, char *buf, int buf_size);
+extern int do_find_trigger_by_query (const char *name, char *buf, int buf_size);
 
 #endif /* _EXECUTE_STATEMENT_H_ */
